+++ changelog.md
@@ -44,6 +44,7 @@
 * Product filter: Option to sort filter results by their display order rather than by number of matches
 * Elmar Shopinfo: Option to export delivery time as availability
 * #654 Place user agreement for downloadable files in checkout process
+* #398 EU law: add 'revocation' form and revocation waiver for ESD
 
 ### Improvements
 * (Perf) Implemented static caches for URL aliases and localized properties. Increases app startup and request speed by up to 30%.